requirements.txt
streamlit>=1.32.0,<2.0
pandas>=2.1.0,<2.3
numpy>=1.26.0,<2.0
scikit-learn>=1.4.0,<1.6
scipy>=1.12.0
joblib>=1.3.0
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0
networkx>=3.2.0
shap>=0.45.0
fpdf2>=2.7.0
requests>=2.31.0
python-dateutil>=2.9.0
packaging>=23.0
tqdm>=4.66.0
